port-config: Disable RGMII_TXC_x autoblock if running at 100MBPS
When running at 100MBPS, the RGMII_TXC_x register in the
Clock Generation Unit must be set to a value which disables 'autoblock'.
This statement contradicts the indications in the SJA1105PQRS Application
Hints Rev. 1.1 document, but has been proven empirically to be true. The
port does not function at 100MBPS if 'autoblock' is enabled.

Issue: ALB-1517
Signed-off-by: Nica Dan <dan.nica@nxp.com>