sja1105p_init: Change loglevel of message about reading an invalid id
Reading a negative device id is not actually an error, as it is expected
to happen in certain situations, and even the driver uses it as a means
of detecting when a reset is complete. Therefore informing the user about
it using KERN_ERR as a loglevel in inappropriate. Changed to KERN_INFO.

Issue: ALB-2890
Signed-off-by: Dan Nica <dan.nica@nxp.com>